fix(ci): resolve 3 failing workflows + 74 CodeQL alerts (#90)
**Workflows fixed:**
- mutation: rename invalid `output-path` to `output` in all stryker configs
- benchmark: absolute artifact paths, explicit JsonExporter, robust file discovery
- docs: exclude Kck.Templates/content from docfx metadata (CPM NU1008)

**Security (CodeQL):**
- Add [Authorize(Roles=Admin)] on Create/Delete endpoints (sample)
- Add user identity check on Delete to prevent IDOR

**Code quality (CodeQL):**
- ElasticsearchSearchService: implement IDisposable for client settings
- Entity<TId>.Id: remove virtual to fix constructor virtual call
- JwtTokenService: replace bare catch with finally pattern
- TotpMfaProvider: add (double) cast to prevent precision loss
- Test fixes: using var for IDisposable, proper HttpMessageHandler disposal

+++ src/Providers/Kck.Security.Totp/TotpMfaProvider.cs
@@ -57,7 +57,7 @@ public sealed class TotpMfaProvider : IMfaProvider
             return Task.FromResult(false);
 
         // Cache used code — TTL = (VerificationWindow + 1) * StepSeconds
-        var ttl = TimeSpan.FromSeconds((_options.VerificationWindow + 1) * _options.StepSeconds);
+        var ttl = TimeSpan.FromSeconds((double)(_options.VerificationWindow + 1) * _options.StepSeconds);
         _replayCache.Set(replayKey, true, ttl);
 
         return Task.FromResult(true);
